Conflicts:
	source/Client/ObjectManager.js
	source/Client/config.default.js
	source/Client/guis/desktop/index.html
	source/Client/guis/desktop/style.css
	source/Server/ObjectManager.js
	source/objects/2.ActiveObject/view.js
	source/objects/3.Exit/common.js
	source/objects/6.Subroom/common.js

+++ source/Client/guis.common/javascript/0.GUI.js
@@ -290,7 +290,18 @@ GUI.initMoveByKeyboard = function() {
 	
 }
 
+GUI.initUndoByKeyboard = function(){
 
+
+    $(document).bind("keydown", function(event) {
+        var ctrlDown = event.ctrlKey||event.metaKey
+        if(ctrlDown && event.which == 90){
+            event.preventDefault();
+
+            Modules.Dispatcher.query("undo", {"userID" : GUI.userid});
+        }
+    });
+}
 
 /**
  * add event handler for removing selected objects by pressing delete-key
@@ -305,16 +316,20 @@ GUI.initObjectDeletionByKeyboard = function() {
 
 				event.preventDefault();
 
-				/* delete selected objects */
-				$.each(ObjectManager.getSelected(), function(key, object) {
+				var result = confirm(GUI.translate('Do you really want to delete the selected objects?'));
 
-					if ($(object.getRepresentation()).data("jActionsheet")) {
-						$(object.getRepresentation()).data("jActionsheet").remove();
-					}
+				if (result) {
+					/* delete selected objects */
+					$.each(ObjectManager.getSelected(), function(key, object) {
 
-					object.deleteIt();
+						if ($(object.getRepresentation()).data("jActionsheet")) {
+							$(object.getRepresentation()).data("jActionsheet").remove();
+						}
 
-				});
+						object.deleteIt();
+
+					});
+				}
 			}
 			
 		}
